% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotBiomasstime}
\alias{plotBiomasstime}
\title{Total biomass plot}
\usage{
plotBiomasstime(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Makes a time series of total biomass for each fish functional group and zooplankton
and benthic resource
}
\details{
X-axis is time. Y-axis is the biomass data in log10 scale.
}
\examples{
sim=simulateFEISTY()
plotBiomasstime(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
P. Daniël van Denderen, Yixin Zhao
}
