% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Food web plot}
\usage{
plotNetwork(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Makes a plot with all feeding interactions of fish functional groups and 
resources.
}
\details{
The round patterns and lines between them represent biomass and feeding fluxes.
The size of the round patterns (biomass) and lines (feeding fluxes) are divided by 
the maximum biomass or flux and re-scaled with the cube root.
The function works on the four prepared setups or revised versions based on these four setups. 
If customized setups by users have more resources and/or fish functional types, this plot function may not work.
In setupBasic and setupBasic2, There is no Y-axis.
In setypVertical and setupVertcal2, Y-axis represent locations of relative surface and bottom to avoid overlap of dot symbols.
}
\examples{
sim=simulateFEISTY()
plotNetwork(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
CCCC, BBBB, DDDD
}
