% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotRates}
\alias{plotRates}
\title{Plots for growth rate, mortality, and feeding level}
\usage{
plotRates(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Plots growth rate (1/year), mortality (1/year), and feeding level (dimensionless) over the size spectrum.
}
\details{
Rate calculation:
\itemize{
\item Averaged values of rate data are estimated over a selected time period 
(default last 40% simulation time).
}
}
\examples{
sim=simulateFEISTY()
# averaged rate data of last 40\%  simulation time
plotRates(sim=sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
P. Daniël van Denderen, Yixin Zhao
}
