% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotSimulation}
\alias{plotSimulation}
\title{Plot FEISTY simulation results}
\usage{
plotSimulation(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
This function is designed to give a quick visualization overview of a simulation output.
}
\details{
It makes a plot combo for the simulation results, including \code{\link{plotSpectra}}, \code{\link{plotRates}},
\code{\link{plotNetwork}} and \code{\link{plotBiomasstime}}.
}
\examples{
sim=simulateFEISTY()
plotSimulation(sim)
# A short cut to call plotSimulation(sim) is provided.
plot(sim)

}
\seealso{
\code{\link{webFEISTY}} A shiny interface for visualizing FEISTY model results \cr
\code{\link{simulateFEISTY}} Run FEISTY model simulations \cr
\code{\link{plotSpectra}} Biomass spectra plot \cr
\code{\link{plotRates}} Plots for growth rate, mortality, and feeding level \cr
\code{\link{plotNetwork}} Food web plot \cr
\code{\link{plotBiomasstime}} Time series of biomass \cr
}
\author{
BBBB, FFFF, DDDD
}
