% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotSSBtime}
\alias{plotSSBtime}
\title{Spawning stock biomass plot}
\usage{
plotSSBtime(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Makes a time series of spawning stock biomass (SSB) for each functional group
}
\details{
X-axis is time. Y-axis is the SSB data in log10 scale.
}
\examples{
sim=simulateFEISTY()
plotSSBtime(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations \cr
\code{\link{calcSSB}} 	Spawning stock biomass calculation
}
\author{
P. Daniël van Denderen, Yixin Zhao
}
