% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotYieldtime}
\alias{plotYieldtime}
\title{Fishing Yield plot}
\usage{
plotYieldtime(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
It displays the time series of the fishing yield for each functional type.
}
\details{
X-axis is time. Y-axis is the fishing yield data in the log10 scale.
}
\examples{
sim=simulateFEISTY()
plotYieldtime(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations \cr
\code{\link{calcYield}} Fishing yield calculation
}
\author{
P. Daniël van Denderen, Yixin Zhao
}
