% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plot.FEISTY}
\alias{plot.FEISTY}
\title{Plot simulation results}
\usage{
plot(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
This function is a shortcut of \code{\link{plotSimulation}}.
}
\details{
See \code{\link{plotSimulation}}.
}
\examples{
sim=simulateFEISTY()
plot(sim)

}
