% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotDiet}
\alias{plotDiet}
\title{Diet plot}
\usage{
plotDiet(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Makes a plot of prey groups in the stomach of each fish functional type and size class. 
The function only works on the four prepared setups or revised versions 
based on these four setups. If customized setups by users have more resources 
and/or fish functional types, this plot function needs to be updated.
}
\details{
To be added.
}
\examples{
sim=simulateFEISTY()
plotDiet(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
Daniel Ottmann Riera, P. Daniël van Denderen, Yixin Zhao
}
