% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Food web plot}
\usage{
plotNetwork(sim)
}
\arguments{
\item{sim}{The dataframe of FEISTY simulation results.}
}
\description{
Makes a plot with all feeding interactions of fish functional groups and 
resources. The size of the biomass circles and feeding fluxes are divided by 
the maximum biomass or flux and re-scaled with the cube root.
The function works on the four prepared setups or revised 
versions based on these four setups. If customized setups by users have 
more resources and/or fish functional types, this plot function may need to be updated.
}
\details{
To be added.
}
\examples{
sim=simulateFEISTY()
plotNetwork(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
Daniel Ottmann Riera, P. Daniël van Denderen, Yixin Zhao
}
