% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Food web plot}
\usage{
plotNetwork(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Makes a plot with all feeding interactions of fish functional types and 
resources.
}
\details{
The solid circles represent biomass of fish size-classes and functional types and the lines between the solid circles represent feeding fluxes.
The size of the circles (biomass) and lines (feeding fluxes) are divided by 
the maximum biomass or flux and re-scaled with the cube root.
The function works on the four prepared setups or revised versions based on these four setups. 
If customized setups by users have more resources and/or fish functional types, this plot function may not work.
In setupBasic and setupBasic2, there is no Y-axis.
In setupVertical and setupVertical2, Y-axis represents the locations of relative surface and bottom to avoid overlap of dot symbols.
}
\examples{
sim=simulateFEISTY()
plotNetwork(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
CCCC, BBBB, DDDD
}
