% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotRates}
\alias{plotRates}
\title{Plots for growth rate, mortality, and feeding level}
\usage{
plotRates(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Plots growth rate (1/year), mortality (1/year), and feeding level (dimensionless) over the size spectrum.
}
\details{
Note the growth rate here denotes the fraction of the available energy can be invested in growth (Eq. 15 in vignette), 
which is not the growth rate from a size class to the next neighboring class (Eq. 16 in vignette).

Mortality panel displays the total mortality, background mortality, fishing mortality and predation mortality. 

Feeding levels calculation is in Eq. 10 in vignettes.

Data are averaged over a selected time period (last 40% of the simulation time).
}
\examples{
sim=simulateFEISTY()
# averaged data of last 40\% of the simulation time
plotRates(sim=sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
BBBB, DDDD
}
