% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEISTY_plotting.R
\name{plotSpectra}
\alias{plotSpectra}
\title{Biomass spectra plot}
\usage{
plotSpectra(sim)
}
\arguments{
\item{sim}{The data frame of FEISTY simulation results.}
}
\description{
Makes a plot of the biomass of all functional groups over the size spectrum.
Data is averaged over the last 40\% simulation time.
}
\details{
The X-axis is individual weight on a log10 scale. The Y-axis is 
biomass on a log10 scale.
}
\examples{
sim=simulateFEISTY()
plotSpectra(sim)

}
\seealso{
\code{\link{simulateFEISTY}} Run FEISTY model simulations
}
\author{
P. Daniël van Denderen, Yixin Zhao
}
